don't add text to the personal nameplate
Name/guild text isn't shown by KNP on the personal nameplate, and
there isn't an option to enable it. I recently got a report asking
for a way to disable it from someone using this. However, it might
be what "your users" are used to? Either way, it's pretty easy to
check for.