Fix errors with profiles lacking FN characteristic
getPlayerdata is being passed a "unit ID" that was being re-mapped
to a "unit ID" by the call to getUnitFullName, which would return
nil and then pass that into getCompleteName which would concat
a nil value with a string and die horribly and loudly at random
points in time.

Now we don't do that, and things are good.